chore: Remove test-only exports
- Remove DownloadReport, TimeTrends, RevenueCircle
- Drop `export` for IndicatorListWrapper and getIcon
  Preserve the one meaningful assertion (router.push on state change)
- Extract ScoreInfo into score-info.tsx
  Port the existing unit tests
