+++ tests/analytics-sidebar-layout.test.tsx
@@ -1,8 +1,5 @@
 import React from 'react';
-import {
-  AnalyticsSideBarLayout,
-  IndicatorListWrapper,
-} from '@/app/[locale]/[state]/analytics/components/analytics-sidebar-layout';
+import { AnalyticsSideBarLayout } from '@/app/[locale]/[state]/analytics/components/analytics-sidebar-layout';
 import { fireEvent, render, screen } from '@testing-library/react';
 
 // Mock opub-ui components
@@ -11,9 +8,7 @@ jest.mock('opub-ui');
 // Mock next/navigation
 const mockPush = jest.fn();
 jest.mock('next/navigation', () => ({
-  useRouter: () => ({
-    push: mockPush,
-  }),
+  useRouter: () => ({ push: mockPush }),
 }));
 
 // Mock components
@@ -86,66 +81,22 @@ describe('AnalyticsSideBarLayout', () => {
     expect(screen.getByText('INDICATORS')).toBeInTheDocument();
     expect(screen.getByTestId('factor-list')).toBeInTheDocument();
   });
-});
 
-describe('IndicatorListWrapper', () => {
-  const mockCurrentState = {
-    code: 'AS',
-    slug: 'assam',
-    name: 'Assam',
-  };
-
-  const mockStatesList = [
-    { code: 'AS', slug: 'assam', name: 'Assam' },
-    { code: 'HP', slug: 'himachal-pradesh', name: 'Himachal Pradesh' },
-  ];
-
-  beforeEach(() => {
-    jest.clearAllMocks();
-  });
-
-  it('renders state selector with correct options', () => {
+  it('navigates to the selected state on change', () => {
     render(
-      <IndicatorListWrapper
+      <AnalyticsSideBarLayout
         currentState={mockCurrentState}
         statesList={mockStatesList}
-      />
-    );
-
-    const stateSelect = screen.getByRole('combobox');
-    expect(stateSelect).toBeInTheDocument();
-
-    // Check if options are rendered
-    expect(screen.getByText('Assam')).toBeInTheDocument();
-    expect(screen.getByText('Himachal Pradesh')).toBeInTheDocument();
-  });
-
-  it('handles state selection change', () => {
-    render(
-      <IndicatorListWrapper
-        currentState={mockCurrentState}
-        statesList={mockStatesList}
-      />
+      >
+        <div>Test Content</div>
+      </AnalyticsSideBarLayout>
     );
 
     const stateSelect = screen.getByRole('combobox');
     fireEvent.change(stateSelect, { target: { value: 'himachal-pradesh' } });
 
     expect(mockPush).toHaveBeenCalledWith(
-      expect.stringContaining('/himachal-pradesh/analytics/')
+      '/himachal-pradesh/analytics/?indicator=risk-score&view=map'
     );
-  });
-
-  it('renders factor list with current state', () => {
-    render(
-      <IndicatorListWrapper
-        currentState={mockCurrentState}
-        statesList={mockStatesList}
-      />
-    );
-
-    const factorList = screen.getByTestId('factor-list');
-    expect(factorList).toBeInTheDocument();
-    expect(factorList).toHaveAttribute('data-state', 'AS');
   });
 });
