--- tests/download-report.test.tsx
@@ -1,126 +1,0 @@
-import React from 'react';
-import { DownloadReport } from '@/app/[locale]/[state]/analytics/components/download-report';
-import { act, fireEvent, render, screen } from '@testing-library/react';
-
-// Mock opub-ui
-jest.mock('opub-ui', () => ({
-  Button: ({ children, onClick, ...props }: any) => (
-    <button onClick={onClick} {...props}>
-      {children}
-    </button>
-  ),
-  Icon: ({ source, ...props }: any) => <span data-testid="icon" {...props} />,
-  Text: ({ children, ...props }: any) => <span {...props}>{children}</span>,
-  Spinner: ({ color }: any) => (
-    <div color={color} data-testid="spinner">
-      spinner
-    </div>
-  ),
-  ShareDialog: ({
-    children,
-    onOpen,
-    onDownload,
-    kind,
-    size,
-    loading,
-    ...props
-  }: any) => (
-    <div
-      data-testid="share-dialog"
-      data-kind={kind}
-      data-size={size}
-      {...props}
-    >
-      <button data-testid="share-dialog-open" onClick={onOpen}>
-        Open
-      </button>
-      <button data-testid="share-dialog-download" onClick={onDownload}>
-        Download
-      </button>
-      {children}
-    </div>
-  ),
-  useScreenshot: () => ({
-    createSvg: jest.fn(),
-    svgToPngURL: jest.fn(),
-    downloadFile: jest.fn(),
-    domToUrl: jest.fn(),
-  }),
-}));
-
-// Mock hooks
-jest.mock('@/hooks/use-window-size', () => ({
-  useWindowSize: () => ({ width: 1024, height: 768 }),
-}));
-
-jest.mock('usehooks-ts', () => ({
-  useMediaQuery: () => true, // Mock as desktop
-}));
-
-// Mock router events
-jest.mock('@/lib/router-events/events', () => ({
-  onComplete: jest.fn(),
-}));
-
-// Mock window.document.querySelector
-const mockQuerySelector = jest.fn(() => ({
-  // Mock DOM element
-}));
-
-describe('DownloadReport', () => {
-  beforeEach(() => {
-    jest.clearAllMocks();
-
-    // Mock document.querySelector without redefining document
-    Object.defineProperty(document, 'querySelector', {
-      value: mockQuerySelector,
-      writable: true,
-    });
-  });
-
-  it('renders download button', () => {
-    render(<DownloadReport />);
-
-    expect(screen.getByTestId('share-dialog-download')).toBeInTheDocument();
-  });
-
-  it('renders ShareDialog with correct props', () => {
-    render(<DownloadReport />);
-
-    const shareDialog = screen.getByTestId('share-dialog');
-    expect(shareDialog).toBeInTheDocument();
-    expect(shareDialog).toHaveAttribute('data-kind', 'secondary');
-    expect(shareDialog).toHaveAttribute('data-size', 'medium');
-  });
-
-  it('handles generateImage function when dialog opens', async () => {
-    render(<DownloadReport />);
-
-    const openButton = screen.getByTestId('share-dialog-open');
-    await act(async () => {
-      fireEvent.click(openButton);
-    });
-
-    // The generateImage function should be called
-    // We can't directly test the async function without more complex mocking
-    // but we can verify the button click works
-    expect(openButton).toBeInTheDocument();
-  });
-
-  it('handles download function when download button is clicked', () => {
-    render(<DownloadReport />);
-
-    const downloadButton = screen.getByTestId('share-dialog-download');
-    fireEvent.click(downloadButton);
-
-    expect(downloadButton).toBeInTheDocument();
-  });
-
-  it('renders Template component with correct structure', () => {
-    render(<DownloadReport />);
-
-    // The Template component is rendered internally
-    // We can verify the ShareDialog is present which uses the Template
-    expect(screen.getByTestId('share-dialog')).toBeInTheDocument();
-  });
-});
