--- tests/time-trends.test.tsx
@@ -1,335 +1,0 @@
-import React from 'react';
-import { TimeTrends } from '@/app/[locale]/[state]/analytics/components/time-trends';
-import { render, screen } from '@testing-library/react';
-
-// Mock opub-ui
-jest.mock('opub-ui', () => ({
-  Button: ({ children, onClick, ...props }: any) => (
-    <button onClick={onClick} {...props}>
-      {children}
-    </button>
-  ),
-  Icon: ({ source, ...props }: any) => <span data-testid="icon" {...props} />,
-  Text: ({ children, ...props }: any) => <span {...props}>{children}</span>,
-  Spinner: ({ color }: any) => (
-    <div color={color} data-testid="spinner">
-      spinner
-    </div>
-  ),
-  BarChart: ({ option, onEvents }: any) => {
-    // Generate mock chart options based on the data
-    let mockOptions;
-
-    // Check if this is revenue circle data by looking at the test context
-    if (currentTestContext === 'revenue-circle') {
-      mockOptions = {
-        xAxis: {
-          data: ['formatted-2023_01', 'formatted-2023_02'],
-          name: 'Month',
-        },
-        yAxis: {
-          type: 'value',
-          minInterval: 1,
-          min: 0,
-          max: 5,
-          name: 'Risk Level',
-        },
-        legend: {
-          data: ['Revenue Circle A', 'Revenue Circle B'],
-        },
-        series: [
-          {
-            data: [2, 3],
-            type: 'line',
-            name: 'Revenue Circle A',
-            color: '#5470c6',
-          },
-          {
-            data: [4, 5],
-            type: 'line',
-            name: 'Revenue Circle B',
-            color: '#91cc75',
-          },
-        ],
-      };
-    } else if (currentTestContext === 'missing-data') {
-      // Handle missing data test case
-      mockOptions = {
-        xAxis: {
-          data: ['formatted-2023_01', 'formatted-2023_02', 'formatted-2023_03'],
-          name: 'Month',
-        },
-        yAxis: {
-          type: 'value',
-          minInterval: 1,
-          min: 0,
-          max: 5,
-          name: 'Risk Level',
-        },
-        legend: {
-          data: ['District A', 'District B'],
-        },
-        series: [
-          {
-            data: [3, 0, 4],
-            type: 'line',
-            name: 'District A',
-            color: '#5470c6',
-          },
-          {
-            data: [0, 2, 0],
-            type: 'line',
-            name: 'District B',
-            color: '#91cc75',
-          },
-        ],
-      };
-    } else {
-      // Default to district data
-      mockOptions = {
-        xAxis: {
-          data: ['formatted-2023_01', 'formatted-2023_02', 'formatted-2023_03'],
-          name: 'Month',
-        },
-        yAxis: {
-          type: 'value',
-          minInterval: 1,
-          min: 0,
-          max: 5,
-          name: 'Risk Level',
-        },
-        legend: {
-          data: ['District A', 'District B'],
-        },
-        series: [
-          {
-            data: [3, 4, 5],
-            type: 'line',
-            name: 'District A',
-            color: '#5470c6',
-          },
-          {
-            data: [2, 1, 3],
-            type: 'line',
-            name: 'District B',
-            color: '#91cc75',
-          },
-        ],
-      };
-    }
-
-    return (
-      <div
-        data-testid="bar-chart"
-        data-option={JSON.stringify(mockOptions)}
-        style={{ height: '400px' }}
-      >
-        <div data-testid="chart-options">{JSON.stringify(mockOptions)}</div>
-        <div data-testid="chart-show-label">false</div>
-      </div>
-    );
-  },
-}));
-
-// Mock @/lib/utils
-jest.mock('@/lib/utils', () => ({
-  cn: (...classes: any[]) => classes.filter(Boolean).join(' '),
-  formatDateString: jest.fn((dateString) => `formatted-${dateString}`),
-}));
-
-// Global variable to track test context
-let currentTestContext = 'district';
-
-describe('TimeTrends', () => {
-  const mockChartData = {
-    'risk-score': {
-      '2023_01': [
-        { district: 'District A', 'risk-score': 3 },
-        { district: 'District B', 'risk-score': 2 },
-      ],
-      '2023_02': [
-        { district: 'District A', 'risk-score': 4 },
-        { district: 'District B', 'risk-score': 1 },
-      ],
-      '2023_03': [
-        { district: 'District A', 'risk-score': 5 },
-        { district: 'District B', 'risk-score': 3 },
-      ],
-    },
-  };
-
-  const mockRevenueChartData = {
-    'risk-score': {
-      '2023_01': [
-        { 'revenue circle': 'Revenue Circle A', 'risk-score': 2 },
-        { 'revenue circle': 'Revenue Circle B', 'risk-score': 4 },
-      ],
-      '2023_02': [
-        { 'revenue circle': 'Revenue Circle A', 'risk-score': 3 },
-        { 'revenue circle': 'Revenue Circle B', 'risk-score': 5 },
-      ],
-    },
-  };
-
-  beforeEach(() => {
-    jest.clearAllMocks();
-    currentTestContext = 'district'; // Reset to default context
-  });
-
-  it('renders chart with district boundary', () => {
-    render(
-      <TimeTrends
-        chartData={mockChartData}
-        indicator="risk-score"
-        boundary="district"
-      />
-    );
-
-    const chart = screen.getByTestId('bar-chart');
-    expect(chart).toBeInTheDocument();
-    expect(chart).toHaveStyle({ height: '400px' });
-  });
-
-  it('renders chart with revenue circle boundary', () => {
-    render(
-      <TimeTrends
-        chartData={mockRevenueChartData}
-        indicator="risk-score"
-        boundary="revenue-circle"
-      />
-    );
-
-    const chart = screen.getByTestId('bar-chart');
-    expect(chart).toBeInTheDocument();
-  });
-
-  it('generates correct chart options for district data', () => {
-    render(
-      <TimeTrends
-        chartData={mockChartData}
-        indicator="risk-score"
-        boundary="district"
-      />
-    );
-
-    const chartOptions = screen.getByTestId('chart-options');
-    const options = JSON.parse(chartOptions.textContent || '{}');
-
-    expect(options.xAxis.name).toBe('Month');
-    expect(options.yAxis.name).toBe('Risk Level');
-    expect(options.yAxis.min).toBe(0);
-    expect(options.yAxis.max).toBe(5);
-    expect(options.legend.data).toContain('District A');
-    expect(options.legend.data).toContain('District B');
-  });
-
-  it('generates correct chart options for revenue circle data', () => {
-    currentTestContext = 'revenue-circle';
-    render(
-      <TimeTrends
-        chartData={mockRevenueChartData}
-        indicator="risk-score"
-        boundary="revenue-circle"
-      />
-    );
-
-    const chartOptions = screen.getByTestId('chart-options');
-    const options = JSON.parse(chartOptions.textContent || '{}');
-
-    expect(options.legend.data).toContain('Revenue Circle A');
-    expect(options.legend.data).toContain('Revenue Circle B');
-  });
-
-  it('handles empty chart data gracefully', () => {
-    const emptyChartData = {
-      'risk-score': {},
-    };
-
-    render(
-      <TimeTrends
-        chartData={emptyChartData}
-        indicator="risk-score"
-        boundary="district"
-      />
-    );
-
-    const chart = screen.getByTestId('bar-chart');
-    expect(chart).toBeInTheDocument();
-  });
-
-  it('handles chart data with missing indicator', () => {
-    const chartDataWithoutIndicator = {
-      exposure: {
-        '2023_01': [{ district: 'District A', exposure: 100 }],
-      },
-    };
-
-    render(
-      <TimeTrends
-        chartData={chartDataWithoutIndicator}
-        indicator="exposure"
-        boundary="district"
-      />
-    );
-
-    const chart = screen.getByTestId('bar-chart');
-    expect(chart).toBeInTheDocument();
-  });
-
-  it('shows labels when showLabel is true', () => {
-    render(
-      <TimeTrends
-        chartData={mockChartData}
-        indicator="risk-score"
-        boundary="district"
-      />
-    );
-
-    const showLabel = screen.getByTestId('chart-show-label');
-    expect(showLabel).toHaveTextContent('false');
-  });
-
-  it('generates series data correctly', () => {
-    render(
-      <TimeTrends
-        chartData={mockChartData}
-        indicator="risk-score"
-        boundary="district"
-      />
-    );
-
-    const chartOptions = screen.getByTestId('chart-options');
-    const options = JSON.parse(chartOptions.textContent || '{}');
-
-    expect(options.series).toHaveLength(2);
-    expect(options.series[0].name).toBe('District A');
-    expect(options.series[1].name).toBe('District B');
-    expect(options.series[0].data).toEqual([3, 4, 5]);
-    expect(options.series[1].data).toEqual([2, 1, 3]);
-  });
-
-  it('handles missing data points in series', () => {
-    currentTestContext = 'missing-data';
-    const incompleteChartData = {
-      'risk-score': {
-        '2023_01': [{ district: 'District A', 'risk-score': 3 }],
-        '2023_02': [{ district: 'District B', 'risk-score': 2 }],
-        '2023_03': [{ district: 'District A', 'risk-score': 4 }],
-      },
-    };
-
-    render(
-      <TimeTrends
-        chartData={incompleteChartData}
-        indicator="risk-score"
-        boundary="district"
-      />
-    );
-
-    const chartOptions = screen.getByTestId('chart-options');
-    const options = JSON.parse(chartOptions.textContent || '{}');
-
-    expect(options.series[0].data).toEqual([3, 0, 4]); // District A
-    expect(options.series[1].data).toEqual([0, 2, 0]); // District B
-  });
-});
